Refactored Code to make Page.tsx smaller
